Some handsome changes after the PR
