+++ front/src/api/model.ts
@@ -1,0 +1,12 @@
+export interface RegisterEntityApi {
+    cardNumber: string;
+    date: string;
+    time: string;
+    firstName: string;
+    lastName: string;
+    idNumber: string,
+    company: string;
+    host: string;
+    signature: string;
+    picture: string;
+}
