+++ front/src/api/routes.ts
@@ -1,0 +1,5 @@
+export const baseRoute: String = "http://localhost:3000/api/registry";
+
+export const backendRoutes = {
+    registerCollection: "/"
+};
