+++ front/src/core/routes.ts
@@ -1,7 +1,5 @@
 import { generatePath } from "react-router";
 
-export const baseRoute: String = "http://localhost:3000/api/registry";
-
 type NavigationFunction = (id: string) => string;
 
 interface BaseRoutes {
